Rename files to singular and fix name conflicts.
